==> 1. ahei.ipynb <==
# AHEI
While looking for a target we found a great candidate named Alternative Halthy Eating Index (AHEI), which scores how healthy is a person diet given its food components (carbo, protein, ...)

## Process
Our team tried to find an already calculated index for measuring diets, but most of them were private, or considered other factors that could introduce a bias on our model, such as underweight or hypertension. In order to fix this, we used the `Global Dietary Database` (GDD) to get each country diet composition.

### Formula Problem
Even tough, the GDD gives us an understanding of the food, we still ran into the problem of calculating it because there were some categories non easily related to our job

import pandas as pd
from sklearn.linear_model import LinearRegression
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error as mse, r2_score as r2
import numpy as np

labels = pd.read_csv("data/raw/GDDLabels.csv", sep=";")
labels = labels.set_index("File").to_dict()["Group"]
labels

ids = ["iso3", "age", "female", "urban", "edu", "year"]

df = pd.DataFrame()
for file, label in labels.items():
    p = pd.read_csv(f"data/raw/NutritionPerCountry/{file}_cnty.csv")
    p = p.query("year == 2018")
    if len(df) == 0:
         df = p[ids]
    df[label] = p["median"]
    print(file)
df

ahei = pd.read_csv("data/raw/AHEI_2018.csv")
ahei = ahei.rename(columns={
    "educ": "edu",
    "median_x": "ahei",
}).drop(columns=["lowerci_95", "upperci_95"])
ahei = ahei.dropna()
ahei

# ahei table is a score of the unique rows and it is difficult to have the same
# person with the same features duplicated in a small "table"
target = ahei.merge(df, on=["iso3", "age", "female", "urban", "edu", "year"])
target

# if nan, then the person may not consume it? red meat for example
ref_target = target.copy().fillna(0)
for col in target.columns:
    if col not in ["iso3", "year"]:
        target[col] = target[col].fillna(ref_target[col].median())
target = target.fillna(0)
target

m = LinearRegression()

x = target[[col for col in target.columns if col not in ["ahei"] + ids]].to_numpy()
y = target["ahei"]
xtrain, xtest, ytrain, ytest = train_test_split(x, y, test_size=0.2)

m.fit(xtrain, ytrain)

print(r2(ytrain, m.predict(xtrain)))
print(r2(ytest, m.predict(xtest)))

from sklearn.tree import DecisionTreeRegressor

t = DecisionTreeRegressor()
t.fit(xtrain, ytrain)

# decision tree nailed it
print(r2(ytrain, t.predict(xtrain)))
print(r2(ytest, t.predict(xtest)))

import pickle

with open("models/ahei_tree.pkl", "wb") as f:
    pickle.dump(t, f)
