==> 2. agg_ahei.ipynb <==
import pickle
import pandas as pd

ids = ["iso3", "age", "female", "urban", "edu", "year"]

labels = pd.read_csv("data/raw/GDDLabels.csv", sep=";")
labels = labels.set_index("File").to_dict()["Group"]
labels

df = pd.DataFrame()
for file, label in labels.items():
    p = pd.read_csv(f"data/raw/NutritionPerCountry/{file}_cnty.csv")
    if len(df) == 0:
         df = p[ids]
    df[label] = p["median"]
    print(file)
df

for col in df.columns:
    if col not in ["iso3", "year"]:
        df[col] = df[col].fillna(df[col].median())
df = df.fillna(0)
df

with open("models/ahei_tree.pkl", "rb") as f:
    m = pickle.load(f)

x = df[[col for col in df.columns if col not in ids]]
df["ahei"] = m.predict(x.to_numpy())
df

agg_ahei = (
    df[["iso3", "year", "ahei"]]
    .groupby(["iso3", "year"])
    .mean()
    .reset_index()
    .sort_values(by=["iso3", "year"])
    .rename(columns={"iso3": "Country Code", "year": "Year", "ahei": "AHEI"})
)
agg_ahei

# creating missing years for interpolation
years = pd.DataFrame({"Year": range(agg_ahei["Year"].min(), agg_ahei["Year"].max()+1)})
countries = agg_ahei[["Country Code"]].drop_duplicates()
years = countries.merge(years, how="cross")
years

full_agg_ahei = (
    agg_ahei.merge(years, how="outer", on=["Country Code", "Year"])
    .sort_values(by=["Country Code", "Year"])
    .reset_index(drop=True)
)
full_agg_ahei

full_agg_ahei = (
    full_agg_ahei.groupby("Country Code", group_keys=False)
    .apply(lambda x: x.interpolate().bfill().ffill())
)
full_agg_ahei

full_agg_ahei.to_parquet("data/processed/interpolated_ahei.parquet")
agg_ahei.to_parquet("data/processed/ahei.parquet")
