==> 3. feature_selection.ipynb <==
import pandas as pd
import numpy as np
from statsmodels.stats.outliers_influence import variance_inflation_factor as vif

# Base

df = pd.read_parquet("data/processed/interpolated_ahei.parquet")
#df.head()

# Prices Database

prices = pd.read_excel("data/raw/prices_for_nutrition.xlsx")
#prices.head()

#prices.tail(10)

prices_calc = prices[:-5]
prices_calc = (
    prices_calc[["Country Code", "Time"] + prices.columns.tolist()[6:]]
    .rename(columns={"Time": "Year"})
    .replace("..", np.nan)
    .interpolate()
    .convert_dtypes()
)
prices_calc

prc = df.merge(prices_calc, on=["Country Code", "Year"])

x = prc.drop(['AHEI', 'Year', 'Country Code'], axis =1)
y = prc['AHEI']

from sklearn.linear_model import LinearRegression
model = LinearRegression()

from sklearn.feature_selection import RFE

rfe = RFE(model, 7)

fit = rfe.fit(x, y)

cols1 = fit.get_support(indices=True)

feat_price = x.iloc[:,cols1].columns.to_list()
feat_price

# CPI

cpi = pd.read_csv("data/raw/cpi.csv")

cpi_calc = (
    cpi.query("(FREQUENCY == 'A') and (MEASURE == 'AGRWTH')")
    [["LOCATION", "SUBJECT", "TIME", "Value"]]
    .rename(columns={"LOCATION": "Country Code", "Value": "CPI", "TIME": "Year"})
    .pivot(columns="SUBJECT", values="CPI", index=["Country Code", "Year"])
    .reset_index()
    .convert_dtypes()
    .sort_values(by=["Country Code", "Year"])
)
cpi_calc["Year"] = cpi_calc["Year"].astype("int")
cpi_calc.head()

# Food and tot may be promissing
cpp = df.merge(cpi_calc, on=["Country Code", "Year"])#.corr()["AHEI"]
#cpp[cpp['Country Code'] == "BRA"].corr()["AHEI"]

cpp = cpp.dropna()

cpp.head(2)

x = cpp.drop(['AHEI', 'Year', 'Country Code'], axis =1)
y = cpp['AHEI']

rfe = RFE(model, 2)

fit = rfe.fit(x, y)

cols2 = fit.get_support(indices=True)

feat_cpi = x.iloc[:,cols2].columns.to_list()
feat_cpi

# WDI

def long_to_wide(df: pd.DataFrame):
    cols = [col for col in df.columns if "20" in col]
    wide = df.rename(columns={"Country Code": "Country"})
    wide = wide.melt(
        id_vars=["Country", "Indicator Name"], value_vars=cols, var_name="Year", value_name="Pct"
    )
    wide = wide.set_index(["Country", "Year"]).pivot(columns="Indicator Name")
    wide.columns = [col[1] for col in wide.columns]
    return wide.reset_index()

wdi = long_to_wide(pd.read_csv("data/raw/world_development_index.csv"))
wdi

wdi_calc = (
    wdi.rename(columns={"Country": "Country Code"})
    .convert_dtypes()
)
wdi_calc["Year"] = wdi_calc["Year"].astype("int")

wdi_rfe = df.merge(wdi_calc, on=["Country Code", "Year"])

x = wdi_rfe.drop(['AHEI', 'Year', 'Country Code'], axis =1)
y = wdi_rfe['AHEI']

rfe = RFE(model, 5)

fit = rfe.fit(x, y)

cols3 = fit.get_support(indices=True)

feat_wdi = x.iloc[:,cols3].columns.to_list()
feat_wdi

# the other features are describes the same thing
feat_wdi = feat_wdi[:1]
feat_wdi

data_final = df.copy()
for feat_df, feat_cols in zip([wdi_rfe, cpi_calc, prices_calc], [feat_wdi, feat_cpi, feat_price]):
    data_final = data_final.merge(
        feat_df[["Country Code", "Year"] + feat_cols], on=["Country Code", "Year"], how="outer"
    )
data_final.columns

data_final.columns = [
    "Country", "Year", "AHEI", "HCI", "CPI", "CPIFoodEnergy",
    "HealthyDietCost", "FruitsShare", "VegetablesShare", "StarchyShare",
    "AnimalSourceShare", "LegumesShare", "OilsShare",
]
data_final

data_final.to_parquet("data/processed/features.parquet")
