==> 5. forecast.ipynb <==
# Forecast
Now we are going to predict the AHEI value given the features we used.
This forecast removes trend and seasonalities to allow the features explaining the target.

import pandas as pd
from prophet import Prophet
from sklearn.metrics import r2_score as r2, mean_absolute_percentage_error as mape

df = pd.read_parquet("data/model_input/features.parquet")

# Predicting countries AHEI
df_pred = df.rename(columns={"Year": "ds", "AHEI": "y"})
regressors = [col for col in df_pred.columns if col not in ["ds", "y", "Country"]]
forecast = pd.DataFrame()
for country, group in df_pred.groupby("Country"):
    m = Prophet(
        yearly_seasonality=False,
        weekly_seasonality=False,
        daily_seasonality=False,
        growth="flat",
    )
    for reg in regressors:
        m.add_regressor(reg)
    m.fit(group)
    country_pred = m.predict(group)
    country_pred["ds"] = group["ds"].tolist()
    country_pred["Country"] = group["Country"].tolist()
    country_pred["y"] = group["y"].tolist()
    forecast = pd.concat([forecast, country_pred])
    print(country)
forecast

# Getting the real part of the dataframe
condition = "ds < 2019"
y, yhat = forecast.query(condition)["y"], forecast.query(condition)["yhat"]

# Checking the metrics value
print(f"r2 {r2(y, yhat)} - mape {mape(y, yhat)}")

forecast.to_parquet("data/model_output/forecast.parquet")
